backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
gunicorn==23.0.0
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
xgboost==2.1.4
lightgbm==4.6.0
imbalanced-learn==0.13.0
joblib==1.4.2
python-dotenv==1.0.1
